% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralSymbODEs.R
\name{coRal.solveCoralWrapper}
\alias{coRal.solveCoralWrapper}
\title{Wrapper function to coRal}
\usage{
coRal.solveCoralWrapper(time, env, pars)
}
\description{
Has the same input and output as run_coral() from the coRal package
}
