% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralSymbODEs.R
\name{defPars}
\alias{defPars}
\title{Default parameters for solveCoral}
\usage{
defPars()
}
\value{
Named list of model parameters.
}
\description{
Helper function, based off of def_pars() from package coRal.
}
\examples{
defPars()
}
