% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralSymbODEs.R
\name{initState}
\alias{initState}
\title{Create initial state for solveCoral}
\usage{
initState(pars, env)
}
\arguments{
\item{pars}{The named list of model parameters.}

\item{env}{The named list of environmental values.}
}
\value{
A named numeric vector containing the initial flux and biomass values.
}
\description{
Helper function, based off of run_coral() from package coRal.
}
\examples{
initState(defPars(), list(L = 30, N = 1e-6, X = 1e-7))
}
\seealso{
\code{\link{solveCoral}}
}
