% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralSymbODEs.R
\name{solveCoral}
\alias{solveCoral}
\title{Solve coral-\emph{Symbiodinium} ODEs}
\usage{
solveCoral(
  times = c(0, 500),
  pars = defPars(),
  env = list(L = 30, N = 1e-06, X = 1e-07),
  lambda = 5,
  method = "vode"
)
}
\arguments{
\item{times}{The time values at which output is desired. Default c(0,500).}

\item{pars}{The paramaters of the model, given as a list. See function \code{\link{defPars}}. Default defPars(1).}

\item{env}{A named list describing the constant environment for the runs. Default list(L = 30, N = 1e-6, X = 1e-7).}

\item{lambda}{The sensitivity of the runs. High values are more sensitive and small values are less sensitive. Default 5.}

\item{method}{The character method argument of ode() from deSolve desired. Default "vode".}
}
\value{
Matrix of values for fluxes, biomass, and host growth rate at explicitly desired time values.
}
\description{
Function solveCoral uses the deSolve package to generate values similar to coRal. Is essentially a transcription of Ferdinand Pfab's Mathematica code to R.
}
\examples{
solveCoral()
solveCoral(times = seq(0,365,0.1), pars = defPars(), env = c(L = 40, N = 1e-7, X = 1e-7), lambda = 10)
}
\seealso{
\code{\link{defPars}}
}
